实现退出登陆后无法直接通过地址访问其它页面的特性

+++ app/Http/Controllers/Administrator/CoursesController.php
@@ -9,15 +9,29 @@ use Illuminate\Support\Facades\DB;
 
 class CoursesController extends Controller
 {
+
     //列出全部学生
-    public function coursesList()
+    public function coursesList(Request $request)
     {
-        $courses = Courses::paginate(16);
-        return view('administrator.coursesList', compact('courses'));
+        //先要验证是否经过登录
+
+        $account = $request->session()->get('account');
+
+        if ($account) {
+            $courses = Courses::paginate(16);
+            return view('administrator.coursesList', compact('courses'));
+        } else {
+            echo '您未登录，请先登录:<a href="/admin/login">重新登录</a>';
+        }
     }
 
     public function editCourses(Request $request)
     {
+        //先要验证是否经过登录
+
+        $account = $request->session()->get('account');
+
+        if ($account) {
         if ($request->method() == 'GET') {
             $id = $request->get('id');
             $courses = Courses::find($id);
@@ -48,17 +62,28 @@ class CoursesController extends Controller
                 echo "<script>alert('编辑失败');location.href='/admin/courses';</script>";
             }
         }
+        }else{
+            echo '您未登录，请先登录:<a href="/admin/login">重新登录</a>';
+        }
     }
 
-    public function deleteCourses(Request $request){
+    public function deleteCourses(Request $request)
+    {
+        //先要验证是否经过登录
 
-        $id=$request->get('id');
-        $courses=Courses::find($id);
-        $res=$courses->delete();
-        if($res){
+        $account = $request->session()->get('account');
+
+        if ($account) {
+        $id = $request->get('id');
+        $courses = Courses::find($id);
+        $res = $courses->delete();
+        if ($res) {
             echo "<script>alert('删除成功');location.href='/admin/courses';</script>";
-        }else{
+        } else {
             echo "<script>alert('删除失败');location.href='/admin/courses';</script>";
+        }
+        }else{
+            echo '您未登录，请先登录:<a href="/admin/login">重新登录</a>';
         }
     }
 }
